Rework Command structure
