+++ Wokkibot/Commands/CommandModule.cs
@@ -1,0 +1,8 @@
+﻿using DSharpPlus.SlashCommands;
+
+namespace Wokkibot.Commands
+{
+    public partial class CommandModule : SlashCommandModule
+    {
+    }
+}
