+++ Wokkibot/Commands/MusicModule/JoinCommand.cs
@@ -1,0 +1,45 @@
+﻿using System.Linq;
+using System.Threading.Tasks;
+using DSharpPlus;
+using DSharpPlus.Entities;
+using DSharpPlus.Lavalink;
+using DSharpPlus.SlashCommands;
+
+namespace Wokkibot.Commands
+{
+    public partial class CommandModule : SlashCommandModule
+    {
+        [SlashCommand("join", "Join your current channel")]
+        public async Task JoinCommand(InteractionContext ctx)
+        {
+            var lava = ctx.Client.GetLavalink();
+            if (!lava.ConnectedNodes.Any())
+            {
+                await ctx.CreateResponseAsync(
+                    InteractionResponseType.ChannelMessageWithSource,
+                    new DiscordInteractionResponseBuilder().WithContent("Lavalink connection is not esbalished")
+                );
+                return;
+            }
+
+            var node = lava.ConnectedNodes.Values.First();
+
+            var channel = ctx.Member.VoiceState.Channel;
+
+            if (channel.Type != ChannelType.Voice)
+            {
+                await ctx.CreateResponseAsync(
+                    InteractionResponseType.ChannelMessageWithSource,
+                    new DiscordInteractionResponseBuilder().WithContent("Channel must be a voice channel")
+                );
+                return;
+            }
+
+            await node.ConnectAsync(channel);
+            await ctx.CreateResponseAsync(
+                InteractionResponseType.ChannelMessageWithSource,
+                new DiscordInteractionResponseBuilder().WithContent("Joined channel")
+            );
+        }
+    }
+}
